noto: move receipt handling to a separate file
